Lands the branch's 68 commits: IVP's OV tree, range manager, ledge tree, object cache, broad phase, pair mindist refresh, pair watcher and default creator, each ported from the disassembly and pinned by calling the shipped vphysics.dll in process; the older IVP ports converted from inches to metres as vphysics runs them (D173); and findings 51's reading of the larger mindist, re-verified. None of it is on the running path yet — that is D172's remaining work.

Gate: TF2DEMOSALVAGE_GCOR_ONLY=1 bash build/gate.sh passed every floor (animation 4686, core 1888, content 1189, rendering 781, scene 737, presentation 454, audio 183, corpus 158, viewer 109, cli 74, logging 17, fonts 7); Viewer3D.UiTests under run-exclusive: 32 passed, 0 failed.

+++ managed/Tf2DemoSalvage.Animation/Animating/IvpContact.cs
@@ -107,7 +107,7 @@ public sealed class IvpContact
         }
 
         // The inverse rotation, which for a unit quaternion is its conjugate.
-        (float X, float Y, float Z, float W) back = (
+        (double X, double Y, double Z, double W) back = (
             -Body.Orientation.X, -Body.Orientation.Y, -Body.Orientation.Z, Body.Orientation.W);
 
         (float X, float Y, float Z) local = IvpQuaternion.Rotate(back, fresh);
@@ -1318,7 +1318,7 @@ public sealed class IvpContact
         }
 
         Quaternion orientation = new(
-            body.Orientation.X, body.Orientation.Y, body.Orientation.Z, body.Orientation.W);
+            (float)body.Orientation.X, (float)body.Orientation.Y, (float)body.Orientation.Z, (float)body.Orientation.W);
 
         Vector3 centre = new((float)body.Position.X, (float)body.Position.Y, (float)body.Position.Z);
 
@@ -1655,7 +1655,7 @@ public sealed class IvpContact
         }
 
         Quaternion orientation = new(
-            body.Orientation.X, body.Orientation.Y, body.Orientation.Z, body.Orientation.W);
+            (float)body.Orientation.X, (float)body.Orientation.Y, (float)body.Orientation.Z, (float)body.Orientation.W);
 
         Vector3 centre = new((float)body.Position.X, (float)body.Position.Y, (float)body.Position.Z);
